Move autoBuildOnSave spec to build-spec.js
