+++ spec/build-spec.js
@@ -688,4 +688,50 @@ describe('Build', function() {
       });
     });
   });
+
+  describe('when the text editor is saved', function() {
+    it('should build when autoBuildOnSave is true', function() {
+      atom.config.set('build.autoBuildOnSave', true);
+
+      fs.writeFileSync(directory + 'Makefile', fs.readFileSync(goodMakefile));
+
+      waitsForPromise(function() {
+        return atom.workspace.open('dummy');
+      });
+
+      runs(function() {
+        var editor = atom.workspace.getActiveTextEditor();
+        editor.save();
+      });
+
+      waitsFor(function() {
+        expect(workspaceElement.querySelector('.build ')).toExist();
+        return workspaceElement.querySelector('.build .title').classList.contains('success');
+      });
+
+      runs(function() {
+        expect(workspaceElement.querySelector('.build')).toExist();
+        expect(workspaceElement.querySelector('.build .output').textContent).toMatch(/Surprising is the passing of time\nbut not so, as the time of passing/);
+      });
+    });
+
+    it('should not build when autoBuildOnSave is false', function() {
+      atom.config.set('build.autoBuildOnSave', false);
+
+      fs.writeFileSync(directory + 'Makefile', fs.readFileSync(goodMakefile));
+
+      waitsForPromise(function() {
+        return atom.workspace.open('dummy');
+      });
+
+      runs(function() {
+        var editor = atom.workspace.getActiveTextEditor();
+        editor.save();
+      });
+
+      runs(function() {
+        expect(workspaceElement.querySelector('.build')).not.toExist();
+      });
+    });
+  });
 });
